+++ packages/browser/src/exports.ts
@@ -60,16 +60,5 @@ export {
 } from './stack-parsers';
 export { eventFromException, eventFromMessage } from './eventbuilder';
 export { createUserFeedbackEnvelope } from './userfeedback';
-export {
-  defaultIntegrations,
-  forceLoad,
-  init,
-  lastEventId,
-  onLoad,
-  showReportDialog,
-  flush,
-  close,
-  wrap,
-  captureUserFeedback,
-} from './sdk';
+export { defaultIntegrations, forceLoad, init, onLoad, showReportDialog, wrap, captureUserFeedback } from './sdk';
 export { GlobalHandlers, TryCatch, Breadcrumbs, LinkedErrors, HttpContext, Dedupe } from './integrations';
