+++ packages/browser/src/index.ts
@@ -35,6 +35,9 @@ export {
   trace,
   makeMultiplexedTransport,
   ModuleMetadata,
+  flush,
+  close,
+  lastEventId,
 } from '@sentry/core';
 export type { SpanStatusType } from '@sentry/core';
 export type { Span } from '@sentry/types';
